+++ _dev/js/theme/core/cart/cartController.js
@@ -3,6 +3,7 @@ import useEvent from '../../components/event/useEvent';
 import submitVoucherHandler from './handler/voucher/submitVoucherHandler';
 import codeLinkSubmitHandler from './handler/voucher/codeLinkSubmitHandler';
 import deleteVoucherHandler from './handler/voucher/deleteVoucherHandler';
+import addToCartHandler from "./handler/cart/addToCartHandler";
 
 const { on } = useEvent();
 
@@ -20,6 +21,7 @@ const cartController = () => {
     on(document, 'submit', '.js-voucher-form', submitVoucherHandler);
     on(document, 'click', discountCode, codeLinkSubmitHandler);
     on(document, 'click', '.js-voucher-delete', deleteVoucherHandler);
+    on(document, 'click', '[data-button-action="add-to-cart"]', addToCartHandler);
   };
 
   return {
