+++ _dev/js/theme/core/cart/index.js
@@ -1,5 +1,4 @@
 import updateCart from '@js/theme/core/cart/updateCart';
-import cartAddProduct from '@js/theme/core/cart/cartAddProduct';
 import cartQuantity from '@js/theme/core/cart/cartQuantity';
 import cartDelete from '@js/theme/core/cart/cartDelete';
 import DOMReady from '@js/theme/utils/DOMReady';
@@ -12,8 +11,4 @@ DOMReady(() => {
   cartQuantity();
   cartDelete();
   init();
-});
-
-$(() => {
-  cartAddProduct();
 });
